Fix Check Game End
