+++ Patches/CheckGameEndPatch.cs
@@ -10,18 +10,6 @@ using static TOHE.Translator;
 
 namespace TOHE;
 
-[HarmonyPatch(typeof(LogicGameFlowHnS), nameof(LogicGameFlowHnS.CheckEndCriteria))]
-class GameEndCheckerForHnS
-{
-    public static bool Prefix()
-    {
-        if (!AmongUsClient.Instance.AmHost) return true;
-
-        if (Options.NoGameEnd.GetBool() && CustomWinnerHolder.WinnerTeam is not CustomWinner.Draw and not CustomWinner.Error) return false;
-
-        return true;
-    }
-}
 [HarmonyPatch(typeof(LogicGameFlowNormal), nameof(LogicGameFlowNormal.CheckEndCriteria))]
 class GameEndCheckerForNormal
 {
